requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
requests==2.32.5
Pillow==11.3.0
pix2text==1.1.6
